Assignment 1: Munros by Train
Assignment 1 for Web & Mobile Mapping by student 2699692. 
Munros are 282 Scottish mountains over 3,000 feet tall, and Munro bagging is a popular outdoor activity in Scotland. There are very few hillwalking maps in Scotland specifically for public transit users, so when trip planning it takes a lot of additional effort to see which mountains are accessible by train because train stations are not usually shown on Munro bagging maps. This map allows users to click on a Munro and see its nearest train station and the distance to that station.